1. Some foods have multiple sizes, able to change in detail page 2. Changed foodlistfragment so that multiple sizes don't show up except in custom search 3. Unavailable data is -1, category is hidden in detailfragment, not added in meal details, not shown in custom search when filter is active 4. Fixed issues with settingsnumberdialog and numberselectdialog due to 'transfat' being a float 5. CustomSearch dialog, by default select nothing when [all]

+++ app/src/main/java/com/example/fastfooddiet/viewcomponent/BarGraph.kt
@@ -16,8 +16,8 @@ class BarGraph(context: Context, attrs: AttributeSet?) : View(context, attrs) {
 
     private var mWidth : Int = 0
     private var mHeight : Int = 0
-    private var value : Int = 0
-    private var limit : Int = 1
+    private var value : Float = 0f
+    private var limit : Float = 1f
     private var valueRect : Rect = Rect()
     private var limitRect : Rect = Rect()
     private var limitEndX : Float = 0f
@@ -54,7 +54,7 @@ class BarGraph(context: Context, attrs: AttributeSet?) : View(context, attrs) {
 
      */
 
-    fun setValues(value : Int, limit : Int) {
+    fun setValues(value : Float, limit : Float) {
         this.value = value
         this.limit = limit
         setDrawables()
@@ -81,7 +81,7 @@ class BarGraph(context: Context, attrs: AttributeSet?) : View(context, attrs) {
     private fun setDrawables() {
         val padding = (0.25f*mHeight.toFloat()).toInt()
 
-        val percent = value.toFloat() / limit.toFloat()
+        val percent = value / limit
         setValuePaint(percent)
 
         /*
